fixed video deforming

+++ main.py
@@ -18,28 +18,32 @@ def send_welcome(message):
 
 @bot.message_handler()
 def function_name(message):
-    streams = []
     try:
         yt = YouTube(message.text)
     except:
         bot.send_message(message.chat.id, 'Invalid URL')
         return
 
-    bot.send_message(message.chat.id, 'Downloading...')
     stream = yt.streams.filter(progressive=True).order_by('resolution').first()
-    filename = uuid.uuid4().hex
-    path = stream.download(output_path="downloads/",
-                           filename=filename)
 
-    video = open(path, 'rb')
-    try:
-        bot.send_document(message.chat.id, video, timeout=1000)
-    except:
-        bot.send_message(message.chat.id, 'File is too large')
+    if (round((stream.filesize / 1024) / 1024) < 50):
+        bot.send_message(message.chat.id, 'Downloading...')
+        filename = uuid.uuid4().hex
+        path = stream.download(output_path="downloads/",
+                               filename=filename)
 
-    os.remove("./downloads/" + filename + ".mp4")
+        video = open(path, 'rb')
+        try:
+            bot.send_video(message.chat.id, video, timeout=1000,
+                           supports_streaming=True, width=640, height=360)
+        except:
+            bot.send_message(message.chat.id, 'Video size is too large')
 
-    bot.send_message(message.chat.id, 'Done')
+        os.remove("./downloads/" + filename + ".mp4")
+
+        bot.send_message(message.chat.id, 'Done')
+    else:
+        bot.send_message(message.chat.id, 'Video size is too large')
 
 
 bot.polling()
